[YS-172] refactor: 스타일 파일 분리 및 불필요한 파일 제거
