+++ src/app/upload/components/CheckboxWithIcon/CheckboxWithIcon.tsx
@@ -2,6 +2,7 @@ import { css, Theme } from '@emotion/react';
 
 import Icon from '@/components/Icon';
 import { colors } from '@/styles/colors';
+import { checkboxLayout, checkboxContainer } from './CheckboxWithIcon.styles';
 
 interface CheckboxWithIconProps {
   checked: boolean;
@@ -47,36 +48,3 @@ const CheckboxWithIcon = ({
 };
 
 export default CheckboxWithIcon;
-
-export const checkboxLayout = (
-  theme: Theme,
-  align: 'left' | 'center' | 'right',
-  size: 'small' | 'large',
-  boldStyle: boolean,
-) => css`
-  ${size === 'small' ? theme.fonts.label.small.M12 : theme.fonts.label.large.R14};
-
-  font-weight: ${boldStyle || size === 'small' ? '500' : '400'};
-
-  color: ${boldStyle ? theme.colors.text06 : theme.colors.text04};
-
-  margin-top: 0.4rem;
-
-  display: flex;
-  justify-content: ${align};
-
-  cursor: pointer;
-
-  label {
-    cursor: pointer;
-  }
-`;
-
-export const checkboxContainer = (size: 'small' | 'large') => css`
-  width: fit-content;
-
-  display: flex;
-  flex-flow: row nowrap;
-  gap: ${size === 'small' ? '0.2rem' : '0.4rem'};
-  align-items: center;
-`;
