+++ src/app/upload/components/InputForm/InputForm.tsx
@@ -1,5 +1,11 @@
-import { css, Theme } from '@emotion/react';
 import React, { ChangeEvent, forwardRef, useState } from 'react';
+import {
+  textInputContainer,
+  textInput,
+  textSubMessageLayout,
+  textCounter,
+  formMessage,
+} from './InputForm.styles';
 
 interface InputFormProps {
   id: string;
@@ -76,54 +82,3 @@ const InputForm = forwardRef<HTMLInputElement, InputFormProps>(
 InputForm.displayName = 'InputForm';
 
 export default InputForm;
-
-const textInputContainer = (size: 'half' | 'full') => css`
-  display: flex;
-  flex-direction: column;
-  gap: 0.4rem;
-  position: relative;
-
-  width: 100%;
-  max-width: ${size === 'half' ? '45.2rem' : '93.6rem'};
-`;
-
-const textCounter = (theme: Theme) => css`
-  ${theme.fonts.label.small.M12};
-  color: ${theme.colors.text02};
-
-  text-align: right;
-`;
-
-const textSubMessageLayout = (showTextCounter: boolean) => css`
-  display: flex;
-  flex-flow: ${showTextCounter ? 'row-reverse nowrap' : 'row nowrap'};
-  justify-content: space-between;
-  align-items: center;
-`;
-
-const formMessage = (theme: Theme) => css`
-  ${theme.fonts.label.small.M12};
-  color: ${theme.colors.textAlert};
-  margin: 0;
-`;
-
-const textInput = (theme: Theme, status: string) => css`
-  ${theme.fonts.label.large.R14};
-
-  width: 100%;
-  height: 4.8rem;
-  padding: 0.8rem 1.2rem;
-  border: 0.1rem solid ${status === 'error' ? theme.colors.textAlert : theme.colors.line01};
-  border-radius: 1.2rem;
-  color: ${theme.colors.text06};
-
-  &:focus {
-    border-color: ${status === 'error' ? theme.colors.textAlert : theme.colors.lineTinted};
-    outline: none;
-  }
-
-  &::placeholder {
-    color: ${theme.colors.text02};
-    ${theme.fonts.label.large.R14};
-  }
-`;
